feat: add reward position ids to earn data props
